backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
google-generativeai>=0.8.0
langchain-text-splitters>=0.0.1
pytest>=7.4.0
pytest-asyncio>=0.23.0
sqlmodel
psycopg2-binary
asyncpg
langgraph>=0.0.10
prometheus-client>=0.20.0
PyYAML>=6.0.1
